fix(typedoc): refactor MapViewProps, MapViewRef & MapViewConfiguration interfaces to fix docu
